# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Ontology Learning API
  version: 1.3.1
  description: >
    Endpoints compatible with your Hamilton parsers (`parse_concept`, `parse_mapping`).
    Sampling behavior:
      - samples=1 → single result form
      - samples>1 → probabilistic lists with likelihood in [0,1]
servers:
  - url: https://api.example.com
...
paths:
...
  /databases/{databaseId}/cluster:
    post:
      tags: [Clustering]
      summary: Start clustering job (async)
      description: >
        Starts a background job to cluster tables. Returns immediately with a job ID.
        Use GET /jobs/{jobId} to check progress and retrieve results.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                applyFinetuning:
                  type: boolean
                  default: false
                  description: If true, apply available finetuned models during clustering.
      responses:
        '202':
          description: Clustering job started
          content:
            application/json:
              schema:
                type: object
                required: [jobId, status]
                properties:
                  jobId: { type: string, example: "job_7f2b1a3c4d5e" }
                  status: { type: string, example: "pending" }
                  message: { type: string, example: "Clustering job started" }
        '404':
          description: Database not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    
    put:
      tags: [Clustering]
      summary: Save updated clustering for a database
      description: >
        Saves the modified clustering result for a database. Use this after users
        have manually adjusted cluster assignments by dragging tables between clusters.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusteringSuggestions'
      responses:
        '200':
          description: Clustering saved successfully
          content:
            application/json:
              schema:
                type: object
                required: [success, message, clustering]
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Clustering saved successfully" }
                  clustering: { $ref: '#/components/schemas/ClusteringSuggestions' }
        '404':
          description: Database not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '400':
          description: Invalid clustering data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  parameters:
    DatabaseId:
      name: databaseId
      in: path
      required: true
      schema: { type: string }
      description: Database identifier returned by the upload endpoint.
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error: { type: string, example: "NotFoundError" }
        message: { type: string, example: "Database not found." }
...
    TableRef:
      type: object
      required: [schema, name]
      properties:
        schema: { type: string, example: public }
        name:   { type: string, example: orders }
    ClusteringGroup:
      type: object
      required: [tables]
      properties:
        label: { type: string, description: Optional human-readable label }
        tables:
          type: array
          items: { $ref: '#/components/schemas/TableRef' }
        scores:
          type: object
          additionalProperties: { type: number }
          description: Optional clustering metrics
    ClusteringSuggestions:
      type: object
      required: [databaseId, groups, createdAt]
      properties:
        databaseId: { type: string }
        createdAt:  { type: string, format: date-time }
        appliedFinetuning:
          type: boolean
          description: True if finetuning was applied for this run
        groups:
          type: array
          items: { $ref: '#/components/schemas/ClusteringGroup' }

    # ---------- Input cores ----------
...